Add SyncCheckpointedReader with implementation of sync and async writeCheckpoints
Summary:
The initial version of CheckpointedReader. In this diff, there is only SyncCheckpointedReader, the next step is to add AsyncCheckpointedReader. SyncCheckpointedReader proxies Reader member functions and inherits CheckpointedReaderBase, which the is class sharing checkpointing functions with both sync and async CheckpointedReaders.

For CheckpointedReaderBase, I've only implemented sync and async checkpoints writing. The next diffs will extend the implementation of CheckpointedReaderBase.

There is still no versioning implemented.

Reviewed By: MohamedBassem

Differential Revision: D17600126

fbshipit-source-id: 6372c6c79360720cf27bf5547f3a89f247f9281e