Fix InternalLogsIntegrationTest.TrimmingUpToDeltaLogReadPointer flakiness
Summary: The test appends a record to snapshot log, then reads snapshot log up to last *released* LSN, then makes a surprised pikachu face if it didn't see the record it just appended. This diff makes it read to tail LSN instead of last released, and simplifies TrimRSMRequest a little.

Reviewed By: lucaspmelo

Differential Revision: D17589838

fbshipit-source-id: 7b0096752e635813f7cf9ae2cb4acd51247852fb