Fix some MaintenanceAPITest flakiness
Summary: It used to set self-initiated rebuilding period to 1 second and expect that it won't get triggered spuriously. It's way too optimistic to expect logdeviced startup to consistently take under a second (especially in dev mode with asan enabled and lots of tests running in parallel). This diff effectively disables self-initiated rebuilding in MaintenanceAPITest, except for the one test case that needs it.

Reviewed By: AhmedSoliman

Differential Revision: D17644568

fbshipit-source-id: 41d29cb0f2de353ea4f88a56e4175728ca9cf4f6