Add nothrow attributes to jemalloc API declarations
Summary:
According to https://github.com/jemalloc/jemalloc/issues/237 and https://fburl.com/aeia78qk, __attribute__((nothrow)) is added to jemalloc API functions to be in compatible with stdlib.h. This diff makes the declarations in `StatsJemalloc.h` to in confirmation with jemalloc declarations.

This problem is discovered while testing fbcode services with the recent version of LLVM compiler. Details of the problem is described in D17679036, which makes the same change to folly.

Reviewed By: tau0

Differential Revision: D17687349

fbshipit-source-id: c3b32f9222fbccad68f8c77ec2fb8975abdf4258