Introducing SocketAdapter
Summary:
As I was transistion code from libevent2 to libevent1. The major task that I
faced was to transition bufferevent. I thought SocketAdapter was the cleanest way of
hiding bufferevent details.
Added bonus is that we will need some similar interface once we move to
AsyncSocket. AsyncSocket can hide under this interface. Idea is that
Connection will own a SocketAdapter. SocketAdapter itself can be of
BufferEventType or AsyncSocketType. We have to choose to initialize one amongst
them depending on a compile
time flag like ifdef LIBEVENT_VERSION. Connection will interface with this
adapter interface and does not need to change if underlying
implementation changes.
SocketAdapter has a minimalistic API to send, receive and manipulate raw
socket, other high level logic is captured in Connection class.
SocketDependencies used extensively for tests will now implement a test version
of SocketAdapter that should help test all the high level handshake, flowgroup,
serialization logic. This should help in fixing existing tests and make sure
they don't break. In future, this helps to mock that interface and test
connection better which is another advantage of this adapter. Lastly testing of SocketAdapater implementation like LibeventSocket
was not done today and will not be done as part of diffs, if anyone feels there
is a need then we could introduce dependencies to particular implementation of
SocketAdapter and test that socketadapter implementation.

Reviewed By: mcrnic

Differential Revision: D15844742

fbshipit-source-id: 9c892f177e09faad76ac7579284a1da7a9400c7b