Configure the admin server to listen on the address in the nodes config
Summary:
Change the listener of the admin API to read the listen address from the nodes configuration instead of the command line arguments. The listen address will be filled by one of two ways:
1- An operator added this node to the config (using addNodes API) or using our expand handlers and also populated its admin address.
2- The node self registered itself with an admin_address that it infered from the command line arguments.

From the server perspective, this means that the `admin-port` and `admin-unix-socket` flag will now only be used for self registration but if self registration is disabled, they won't play any role in influencing the listen address.

Standalone admin server on the other hand will continue to infer its listen address from the command line argument.

Given that it's a requirement now to have `admin_address` in the nodes config of the node if the admin server is enabled, I started populating this field in the integration test utils.

Reviewed By: AhmedSoliman

Differential Revision: D17475139

fbshipit-source-id: 49fc14a51ddd6889c388b43b2003813aee69e848