MaintenanceManager: Remove maintenances once they are empty
Summary: When all the nodes in a certain maintenance get removed from the cluster, there's no point of keeping this maintenance anymore. Let's asynchronously write a delta to remove it. To prevent trying to remove the same maintenance multiple times, we will only one removal delta in flight at the same time.

Reviewed By: AhmedSoliman

Differential Revision: D17761082

fbshipit-source-id: ac6ba82338beb2f703589af4c23bf855419d935d