Making Socket::isClosed mirror bev_
Summary:
bev_ ptr was used widely in Socket to detect whether a connect attempt is in progress or not. It also meant whether a socket was closed or not.

Once we use SocketAdapter in Connection bev_ won't be initialized . We will remove bev eventually but untill then we need to have a way to replicate what bev_ stood for and this should work for both Legacy connections and new asyncsocket based connections.

I am introducing conn_closed_ which is basically a shared_ptr<bool> which is initialized when a connection is attempted in case of client socket or when connection is created in case of server sockets.

On closing the connection we set this flag to true and it basically behaves/mirrors bev_ ptr. This makes it possible to keep all the existing logic in Socket.cpp as is and maintain invariants.

Reviewed By: mcrnic

Differential Revision: D16954562

fbshipit-source-id: 27fe4b8d77011ecf3a6da328767bb140b16b0114