Move the versioning implementation from CheckpointedReaderBase to CheckpointStore
Summary:
Moved the handling versioning logic to CheckpointStore by changing the UpdateCallback to StatusCallback. The UpdateCallback takes version and value from the store in case of version mismatch, so the version handling should be implemented there.
So far the UpdateCallback was in the CheckpointedReaderBase but this class should not implement it.

Reviewed By: MohamedBassem

Differential Revision: D17876948

fbshipit-source-id: 813a8aa5643bd833070dd13307e8c8c0c1bbbf05