MessageReader to read messages from AsyncSocket
Summary:
Read path for async socket.
Checkout the comment in MessageReader.h to understand how the read callback works. I have a feeling this can be made cleaner with a moveable buffer. Also I think instead of allocating a buffer everytime we can do something better by allocating a big buffer and using that for a while till we exhaust this buffer. I will create both of these as followups.

Reviewed By: mcrnic

Differential Revision: D15920197

fbshipit-source-id: a1787e286f9a081e1af23eb8551a033e4e7b3f37