Use async socket API for sending buffers to socket.
Summary:
All the serialized data was already getting written to a IOBuf in a previous diff. Now we have to use this IOBuf and write it into asyncSocket.

Added a interface sendBuffer that is overridden in Connection to write into SockAdapter. Default implementation adds the buffer to evbuffer by copy.

Reviewed By: mcrnic

Differential Revision: D15886783

fbshipit-source-id: 471b5a9922bb1f7331d416674b8d5d5c3d9197f4