Introducing ProtocolHandler and IProtocolHandler
Summary:
In the final model, this class is the bridge between raw AsyncSocket and Connection. It shuttles between Connection that executes on some thread in CPU threadpool and raw socket that exists some network thread pool.

For now I am using this class on reverse path from AsyncSocket to Connection. It helps to forward, errors, data and other notifications from AsyncSocket to Connection.

In future, this class will wrap AsyncSocket  and will have api to connect, write into the socket, etc. The Connection will invoke ProtocolHandler API and
internally it will enqueue request to the right network thread to send the data out.

On reverse path this class will forward messages read by SocketReadCallback and forward it to message specific handlers

Reviewed By: mcrnic

Differential Revision: D17197507

fbshipit-source-id: 2ee6862922570d8a70bc1e951c1e803fb8959186