LOGDEVICE_TEST_LEAVE_DATA_IF_FAILED environment variable
Summary:
Like `LOGDEVICE_TEST_LEAVE_DATA`, but only leaves data if gtest reports a failure. Should be useful for flaky tests.

Couldn't use `folly::test::TemporaryDirectory` because it requires you to say *in advance*, before creating directory, whether or not the directory will need to be kept. (This would be easy to change in folly, but I just don't want to go through folly review process, and we would still have to wrap the folly TemporaryDirectory into something that checks the environment variable on destruction.)

Reviewed By: tau0

Differential Revision: D17893003

fbshipit-source-id: 1b03773c47be04de91b646f23f51a06f7fd02fff