Fix some derpy slowness in settings initialization
Summary:
`UpdateableSettingsBase::setDefaultValue(name, value)` is a method that parses the default value of the given setting from the given string and saves it. To do that, it creates and uses a temporary `boost::program_options::options_description` that contains description of... *all* settings, for some reason. This makes settings initialization O(n^2) because it calls setDefaultValue() for each setting, which then also iterates over all settings to add then to `options_description`. I couldn't figure out why it's done this way instead of creating `options_description` with only one setting.

Forgivable for hackathon code.

Reviewed By: mcrnic

Differential Revision: D17890183

fbshipit-source-id: 149cf31bd2e688747febd47f31043b138eb11bc5