Socket health detection
Summary:
Slow socket detection to resolve issues seen in gorilla and maybe silently seen in other global connections.

The exact problem and resolutions suggested are here in the post : https://fb.workplace.com/groups/netstigations/permalink/2821652817883315/

I am attempting the closing of low throughput network congested socket  so that hopefully it picks up a better route next time. For this, I am collecting throughput stats for the socket and if the socket throughput drops below some configurable expected value, and TCP_INFO indicates that it was network congestion which caused this low throughput. The socket is closed.

In order to avoid unnecessary closure of idle sockets or low rtt sockets, sockets are classified as active and idle. A socket is considered active if it has greater than or equal to threshold amount of bytes pending in evbuffer. Moreover to qualify the socket as active, it needs to have those many bytes for a threshold amount of time in the past socket health check period. In all other cases the socket is considered idle and has throughput of 0.

For simplicity, throughput calculations are done based on bytes added in evbuffer. Also it is assumed that when onBytesPassedToTCP is invoked, those bytes are delivered to the remote end.

In order to allow conditions to fix by themselves and avoid socket creation churn, socket closures are rate limited. The rate limit is configurable.

Reviewed By: al13n321

Differential Revision: D17589467

fbshipit-source-id: 0c49b54042b087f0488e6043d6a0b8f197d6753e