Pass set of shards that have already passed safety check to safetycheck scheduler
Summary: With this diff we will pass a ShardSet to Safety check scheduler which contains the shards that have an active workflow and whose storage state is one of RW_TO_RO, RO, DATA_MIGRATION. This set essentially indicates the set of shards that have either passed safety check earlier or shards that have skipped safety check and are inaccessible. The SafetyCheckScheduler will make use of this set such that these shards are considered as DISABLED for the purposes of current safety check. This fixes the case where previously a sequence of MAY_DISAPPEAR maintenances will pass the check because the previous safety check results did not impact the next safety check result.

Reviewed By: MohamedBassem

Differential Revision: D18047616

fbshipit-source-id: ecb551d67a7bc3e8181b824f431409ef8f418fbc