Make CheckpointStateMachine a generic KeyValueStoreStateMachine
Summary:
In the first version, we had CheckpointStateMachine, which was maintaining the map from customer id to their checkpoint.

We decided to make it more generic, so that it will maintain the map from string to string. We call it KeyValueStoreStateMachine. In particular, the checkpoints can be also maintained there, but in serialized form.

In this diff, I moved the previously created code to common/, renamed classes/structures/functions and updated unit tests to match the generic version of the StateMachine.

Reviewed By: MohamedBassem

Differential Revision: D18059188

fbshipit-source-id: 652c02227bcac1223554d67394c517f77bd8a9cb