Disable recycling server sockets.
Summary:
Socket instance in case of server socket was long lived. Once created only in few cases it was deleted and recreated. On connection errors, or various other cases the bufferevent within the Socket instance was recreated on Socket::connect invocation.

With AsyncSocket similar recreation of socket on errors is possible. But to do that I would have to pass in instance of some SockAdapterFactory to Connection and use that to recreate the sock_adapter when required. Instead of doing I felt saving the socket instance was not that useful anyway(Please review if I missed something here). Hence in this diff I just delete the connection instance everytime a server socket is closed for whatever reason.

connect_throttle was once that needed to stay remain valid across multiple server socket reincarnations. I moved it out of the connection and saved it in LibeventCompatibilityConnectionFactory. This same connectThrottle instance could be used for various Connection instance to same server endpoint.

Reviewed By: tau0

Differential Revision: D17065460

fbshipit-source-id: 69d94929cd7783c2a1ba288af72ff4250eed3a2b