Introduce AsyncSocketConnectionFactory and LibeventCompatibilityConnectionFactory
Summary:
AsyncSocketConnectionFactory creates a connection using the AsyncSocketAdapter.

LibeventCompatibilityConnectionFactory creates a connection using AsyncSocketAdapter factory or LegacySocketAdapter factory depending on what evbase type was use to instantiate. If LibeventCompatibilityConnectionFactory is instantiated with EvBase of LEGACY type all connections created by this instance will be of with legacy socket adapter. Similarly to create AsyncSocket based connections LibeventCompatibilityConnectionFactory is  instantiated with EvBaseType FOLLY.

An instance of LibeventCompatibilityConnectionFactory is owned by sender and it replaces the current LegacyConnectionFactory in Sender.

Reviewed By: mcrnic

Differential Revision: D16445177

fbshipit-source-id: 78f54c57d9d7ac44df03c2aa9381178c31a24358