Introduce switch in socket to create bufferevent or not.
Summary:
Socket.cpp needs to support bufferevent and also has logic for connection with asyncsocket. It takes decision to create the bufferevent on incoming connection.

Other promising option was to add an argument to the constructor to indicate whether to create bufferevent. If you prefer that  I can change this implementation to add that argument.

Reviewed By: mcrnic

Differential Revision: D17552505

fbshipit-source-id: cbb4637d1d076a055ebaaf4cb863bf7d3af86137