Add sender shutdown sockets and Add parameterized tests for asyncsocket and legacy socket.
Summary:
Shutdown of AsyncSockets has to happen on Eventbase threads on which they were created. Current shutdown process destroys sender on main thread which fails this AsyncSocket expectation.

I think I broke this expectation when I severed the inheritance relation between EventLoop and Worker. Before that diff, we used to join the EventLoop thread which destroyed the Worker and all the sockets held in it's Sender.

Fixing it here but posting and waiting for sockets to close as part of Processor::shutdown.

Reviewed By: mcrnic

Differential Revision: D17552999

fbshipit-source-id: d1a65a5445748375d9f83a7203fcf0cda963be01