Add writeWithoutCopy to ProtocolWriter
Summary:
writeWithoutCopy within ProtocolWriter was creating a weak reference on payload
buffer. Because of this the owner of the buffer had to stay around until the
buffer is written into the socket. But in the new model we call onSent
immediately because of which the buffer could be deleted before getting copied
into the socket.

This new api accepts a strong reference of a iobuf chain which can be cloned
and hence the underlying buffer will stay valid untill it gets written into the
socket.

Reviewed By: mcrnic

Differential Revision: D18075278

fbshipit-source-id: 0d2e466097efe6ff48d03eea59d08af250b399f6