Copy record buffer from storage into IOBuf
Summary:
Record buffer from storage is copied and moved into the Record_Message instance which is ultimately sent on network.
This instance manages the lifetime of the buffer. In new model, socket on sent
will be called immediately, even before writing into the socket hence the
lifetime of underlying buffer needs to be extended.

This diff uses IOBuf to do just that. The buffer is cloned
when it is added to the serialized buffer.

Reviewed By: mcrnic

Differential Revision: D18074871

fbshipit-source-id: a1bbfe18f2e2fc995fb021ed11d8c8b6b7f18db3