Fix checkSocketHealth for new connection.
Summary:
CheckSocketHealth needed implementation for the new connection. The number of bytes buffered logic changed in the new connection. Hence we needed to use the new api to get that to detect whether we are above idle-threshold.

Also onBytesPassedToTCP is not called for the new connection, hence we have to moved the num_bytes drained logic into Connection.cpp.

This diff address both these issues.

Reviewed By: mcrnic

Differential Revision: D18078433

fbshipit-source-id: f1a19b83ed8d5f8221bbf24d117e6194d18ec0c9