PayloadHolder is now a wrapper for folly::IOBuf
Summary:
PayloadHolder is used to manage owned and unowned payload buffers. But it was
designed with evbuffers in mind. Now I have made serialization based on iobuf
and also introduced iobuf in PayloadHolder. After these changes, it is
unnecessary to have PayloadHolder need to track, owner information or flat
payload. All these features are provided by folly::IOBuf. PayloadHolder is used
everywhere hence not replacing PayloadHolder with folly::IOBuf for now. Also, PayloadHolder
provides Payload objects which are weak references to the underlying buffer.
For this reason as well I cannot remove PayloadHolder for now easily.

TailRecord is also made to use ProtocolWriter::writeWithoutCopy in this diff.

Reviewed By: mcrnic

Differential Revision: D18055607

fbshipit-source-id: 6eecff015416305ba45de464ef271e37dab42f31