Make ReplicatedStateMachine own its SyncSequencerRequest-s instead of firing-and-forgetting
Summary:
A little refactoring.

This will allow the next diff in the stack to stop a SyncSequencerRequest when it wants rather than when callback is called. It also seems like a nicer pattern in general: if the "parent" state machine owns its "child" state machine and has a pointer to it, it's (a) more observable: easy to see the state of child state machines when inspecting the parent (e.g. with admin command or gdb), less hidden state, (b) easier to shut down: the parent can just synchronously stop the child. The downside here is that the SyncSequencerRequest now lives on the RSM's worker rather than worker hash(log_id)%num_workers, so the GetSeqStateRequest merging won't apply to it; that seems fine since there aren't many RSMs, and they don't do a lot of GetSeqStateRequest-s.

Reviewed By: pvjykumar

Differential Revision: D17777348

fbshipit-source-id: 810f416c0599e5e69521d6df981ff40d74a25634