Fix logdeviced crash if ServerParams initialization fails. Fixes #133
Summary: If an exception happens during the initialization of ServerParmeters, we do an ld_critical which invokes bumpErrorCounter that uses the StatsHolder of the server, but the Server is not there anymore so we crash. This case is mostly handled in the destructor of ServerParameters but it doesn't get called because of the exception. This diff does the init logic in an init() function instead of the constructor.

Differential Revision: D18138159

fbshipit-source-id: 042173089295be4ae566f821d784c2d0c4cff187