Fix dispatchMessageBody to check if socket adapter is valid.
Summary:
Connection works with both old and new connections. In case of old connections
socket adapter is null.

Connection::dispatchMessageBody was accessing socket adapter when hit with
ENOBUFS.
The new socket adapter code is disabled and we had tests for this scenario but
because the tests created Socket instead of instance of Connection we missed
this test case.

This can be resolved by setting should-be-inlined but to avoid this for good we
should land this immediately.

Differential Revision: D18211490

fbshipit-source-id: e9e67917789b32a19ae0dc9a82f59a51f58e5028