Make NodeStatsControllerIntegrationTest mock successful appends as well as unsuccessful
Summary:
Before this diff, the test injects errors on client side for appends that it wants to fail, but for appends that it wants to succeed it goes through the full append path, through sequencer and storage nodes and all. This was causing two problems:
 1. The real appends sometimes randomly fail under load, and the boycotting setup in this test is too sensitive to that. In real life it would be better because there would be much more nodes and more appends. (Would be better to use a unit test instead of integration test for that - then we can have realistic number of nodes and append stats.)
 2. To pretend that a sequencer node is bad, the test sets sequencer affinity for some log to that node and injects failures for that log. When the node gets boycotted, sequencer moves elsewhere, and that other node starts accumulating bad karma from injected failures. Sometimes that causes the second node to get boycotted and break the test.

This diff fixes it by changing how error injection works to allow injecting successes and deciding based on node id rather than log id.

Reviewed By: tau0

Differential Revision: D18055978

fbshipit-source-id: 7a685f9602097e8812b82599b0d1890a1ceb694d