Client fix to handle ENOBUFS without losing the message.
Summary:
We dropped a message body when we hit ENOBUFS when dispatchingMessage to the worker. This screwed up the socket stream as we read the next bytes in evbuffer which was next message's protocol header. This lead to kepler client failing ld_check failures or closing connections with bad message and in the worst conditions pushing a record to be processed which lead checksum failures on the records.

This lead to parity stuck readers.

The solution here is to not drain the evbuffer from the input stream before passing it to worker but after passing it to worker or processing the message.

If the dispatch fails with NOBUFS we schedule a zero timeout and retry dispatch at the next eventloop. If dispatch succeeds  we drain the bytes from the evbuffer.

This diff should address NOBUFS scenario and also inline_message_execution. Reviewers should take care that is the case.

Reviewed By: mcrnic

Differential Revision: D18278747

fbshipit-source-id: 6a54a343079e9ff6918b628438c92f37ea7491a6