Fix data race in RSMBasedVersionedConfigStore
Summary:
The main thread calls the constructor of RSMBasedVersionedConfigStore which creates the underlying RSM. However, the RSM needs to read all its logs to be ready to use,  so the CTOR doesn't block on it, but passes a callback to RSM. At some point, the callback is called from different thread (the worker thread, because I'm using Processor to post the requests).

So on the main thread, construction of VCS is done but the RSM is not ready to use.
At this point, the main thread may call getConfig on VCS, but if the RSM is not ready, it returns E::AGAIN, but we need to check the variable ready_ first.

So in general, there may be at most 2 threads using the VCS fields:
- the main thread, which calls the constructor and all the member functions (there shouldn't be multiple threads calling them)
- the RSM thread which calls the callback created in CTOR.

But every field that is used in the callback is already protected by either a mutex (state_) or is atomic (ready_). Other fields (extract_fn_ pointer, processor_, etc. ) are not used in the callback, so at most 1 thread will use them, so they are not protected.

The tsan stack trace generated before the fix: P121286242, generated with the command:

`buck build mode/dev-tsan logdevice/test:tests && time seq 5 | parallel --gnu -j50 'buck-out/gen/logdevice/test/tests --gtest_filter=RSMBasedVersionedConfigStoreIntegrationTest.UpdatesAndGetsAfterRestart > ~/t/d/{} 2>&1 || (echo {} failed && cat ~/t/d/{} | ./experimental/tau0/bash_tooling/symbolize.py buck-out/gen/logdevice/test/tests > ~/t/d/{}.symb)'`

After the fix, the tests don't fail.

Reviewed By: tau0

Differential Revision: D18324232

fbshipit-source-id: 9b70317fdabda2090792fb3c0cfd9f06f558b420