Fix race condition in UpdateableSecurityInfo
Summary:
It was outright accessing a field from different threads without synchronization.

This diff refactors UpdateableSecurityInfo to put all its mutable state in one struct behind an UpdateableSharedPtr, instead of having multiple UpdateableSharedPtr-s for different fields. This seems simpler and prevents torn reads (which probably aren't really a concern, but why not).

And also along the way adds UpdateableConfigBase::callAndSubscribeToUpdates(), similar to existing UpdateableSettings::callAndSubscribeToUpdates(), to prevent a potential issue with onConfigUpdate() racing against itself when called from UpdateableSecurityInfo constructor and from config update thread at the same time (racing would be ok, except that the instance with older view of the config may win the race, and an update will be lost).

And removes some unnecessary logging on startup.

Reviewed By: davidvigier

Differential Revision: D18287100

fbshipit-source-id: 7580bbd4000f77f9744c8ffbb84e8b13148533fe