Small optimization to drop socket early.
Summary:
There is a delay between socket going bad( ProtocolHandler::notifyErrorOnSocket ) and closing the socket. The delay is next eventloop iteration.

Socket::sendBuffer still accepts the message and fakes that the message was sent. We cannot call onSent(E::INTERNAL) or something like that before of asserts in onSent for various state machines.

This is not useful as the message will be retried by the state machine after a timeout for the message_reply which can be long. This situation worsens when AsyncSocket moves to another thread. The delay between socket going bad and socket getting closed might be larger than one eventloop iteration the solution to that will  be to post into the worker HI_PRI hence polluting the HI_PRI queue.

Sender already checks if the socket isClosed before using it. Introducing method good(). This is used by sender to know whether the socket is still good and avoid pushing messages in a bad socket thus reducing the delay to actually send out the messages .

Reviewed By: tau0

Differential Revision: D18296234

fbshipit-source-id: 4fd201f11d08bdf7055c5aa90e46b5930d2b2122