Fix UnreleasedRecordDetectorTest flakiness caused by connection throttling
Summary: The test creates and starts a Processor using a config pointing to some unix domain socket, then starts listener on that socket. The Processor may decide to do some random background work and try to connect to the node before the listener starts. The connection will fail, and by default ConnectThrottle would prevent reconnection attempts for the next milliseconds, which is enough to fail an append later in the test. This diff disables connection throttling. I also considered deferring Processor startup until listener starts, but epoch store initialization needs Processor to be running, so it wouldn't be trivial.

Reviewed By: davidvigier

Differential Revision: D18372600

fbshipit-source-id: f2057714d79f367015457c1d326b22a32ebbf2f9