Reintroduce scheduled writes instead of invoking onSent immediately
Summary:
This diff reintroduces scheduled writes instead of invoking onSent immediately
somewhat like in case of legacy sockets.

Our initial plan was to invoke onSent immediately once we add a message to AsyncSocket. But that does not work as our state machines have been written such that onSent is always invoked asynchronously when the message is sent successfully.

My first attempt to deferring onSent is here D18296187, this is my second attempt which has been working fine and passed all the e2e tests for hdd and flash clusters without perf regressions.

More about this diff :
1. Added a EvTimer in Connection , schedule_write_chain_ to schedule the message instead of using the worker. This was done as a optimization and could be separated from this diff let me know if you want that. Earlier mechanism was posting a HI_PRI message on the worker, this pollutes the HI_PRI queue with socket related stuff, HI_PRI queue can be delayed in case of bursts we do not want to delay messages because of that.
2. Rest of the changes are related to actual problem.
- In case of legacy sockets we had a method onBytesPassedToTCP which was invoked when libevent added stuff from  evbuffer into the tcp socket. We now invoke onBytesPassedToTCP when we invoke writeChain on asyncSocket not exactly similar to older method but it satisfies all the expectations for state machines above it.

- Introduced onBytesAdmittedToSend which is similar older onBytesPassedToTCP. New onBytesPassedToTCP update sender level stats and update socket health stats.

The reason for doing this is correctly calculate getBytesPending(). getBufferedBytesSize() also was added for the same reason. In case of legacy sockets, we added a message into sendq and into evbuffer, Socket::getBytesPending was sum of pendingq, serializeq, sendq and bytes buffered in evbuffer. Sender::getBytesPending was a summation of all Socket::getBytesPending, but instead of the costly operation of going over all the sockets everytime  as this is checked for every message, what we do is we have Sender::noteBytesQueued and Sender::noteBytesDrained, when a message is added into Socket we call notesBytesQueued once during registerMessage and once after message is added to evbuffer.
noteBytesDrained is called when Message::onSent is called and during onBytesPassedToTCP.

But now with these changes, message will be in Socket::sendq_ or the serialized iobuf will be in Connection::sendChain_ till we flush the sendChain_ into the AsyncSocket. At that point, AsyncSocket can flush into the tcp socket but if socket is slow or ingest into the socket is high it will add this request into its own internal queue. Hence, if I invoke onBytesPassedToTCP after adding into asyncSocket and drain those bytes from sender we skip over bytes buffered into asyncsocket and that makes getBytesPending at sender incorrect. It also breaks socket health detector, which depends on bytes buffered in socket to determine whether socket is idle or not idle. Hence, I introduced getBytesBufferedSize, this is amount of bytes under Connection layer, in case of legacy sockets its just drain_pos_ - next_pos_ and in case of new AsyncSocket its drain_pos_ - next_pos_ to account for serialized bytes in sendChain_  plus actual bytes in asyncsocket given by socket_write_cb_.bufferedBytes().

Added stats that provide details on schedule delay in writes.

Reviewed By: mcrnic, tau0

Differential Revision: D18142836

fbshipit-source-id: b28060c3bd5fd6b263336824b2e041bdeab30ef1