Fix use-after-free in Listener shutdown
Summary:
There was a use-after-free from a lambda enqueued in folly::AsyncServerSocket::RemoteAcceptor::stop(), trying to call acceptStopped() on a logdevice::ConnectionListener that was destroyed. That's because, when Listener added itself as a listener using AsyncServerSocket::addAcceptCallback(), it provided a non-null event base, which caused AsyncServerSocket::stopAccepting() to call acceptStopped() asynchronously, after acceptStopped() returns; while ~Listener() expected no callbacks to be called after acceptStopped().

This diff trivially fixes it by telling AsyncServerSocket that callbacks should be called on the same event base on which the socket lives (which was already the case, but AsyncServerSocket doesn't check for equality of the two provided event bases).

Reviewed By: gdrane

Differential Revision: D18372920

fbshipit-source-id: 2c98c4f1b57d8fe504747e225a7cfa663cdeeffb