Fix assert in LogRecoveryRequest::onSealReply()
Summary: Assert checked that if LogRecoveryRequest's completion condition is reached, LogRecoveryRequest should have been destroyed. But LogRecoveryRequest destruction is not immediate, it uses a timer with zero delay, so the assert may squeeze between completion condition becoming true and LogRecoveryRequest being destroyed. This diff checks if the deferred destruction timer is active in addition to the completion condition.

Reviewed By: runemaster

Differential Revision: D18425664

fbshipit-source-id: ace0e83c408c95f35538b8b4df18128f085e6749