Reintroduce scheduled writes instead of invoking onSent immediately
Summary:
UPDATE: Found a leak in sender::bytes_pending_ stat during socket close for
asyncsocket(D18386579) here are the changes. Fixed that and added tests to make sure bytes_pending reach to
zero at shutdown.

=========================== Previous summary landed as part of D18142836 and then backed out because of the bug=========================
This diff reintroduces scheduled writes instead of invoking onSent immediately
like in case of legacy sockets.

One problem was that if the socket is bad Appender could not handle the error.
We have broken a assumption in previous half that messages can still be
received after socket is closed. RecoveryNode does not handle receiving
messages after socket closure. Interestingly the because of some timing related
thing this problem  does not happen when onSent is delayed. Will have to debug
more because we are just hidden from this problem as of now.

Also, added good method to Connection which indicates if socket is useful.
Sender removes the socket from the map immediately if it finds that the socket
is not useful.

Added stats that provide details on schedule delay in writes.

Finally, fixed a serious bug which can happen if serialized size of a message
is not same as current protocol len. I was incrementing notesBytesQueued in
socket by msglen and decrementing in onBytesPassedToTCP by serialized len.
Serialized len can be different from msglen if the endpoint is running on a
different protocol.

Reviewed By: mcrnic

Differential Revision: D18410678

fbshipit-source-id: 8ef11e8328bd26e068531fb04d1b70a633efdff4