Fix confusing error message when a shard directory is missing
Summary:
Without this diff, when one of the directories /data/logdevice/4440/shard* doesn't exist, we print a very unclear message:
  ShardToPathMapping.cpp:167] assignFreeSlots() Found 1 unassigned shards but 0 free slots under "/data/logdevice/4440"!

That's because ShardToPathMapping supports mixing two styles of shard paths: `/shard*/` and `/slot*/shard*/`, and when some shards don't exist in one style it falls back to the other style. Slots are an old feature that we probably never used, and by now probably no one remembers that it exists.

This diff makes ShardToPathMapping avoid mixing the two styles: it's either all `/shard*/` or all `/slot*/shard*/`. If we find both, error out.

Reviewed By: tau0

Differential Revision: D14741769

fbshipit-source-id: 8f4c446c75ed7e757cd77e444b1da0f9e9519b22