Add SteadyTimestamp::toString()
Summary:
Timestamp.h tries to be clever and insists that SteadyTimestamp shouldn't be too easily convertible to string because such conversion is nondeterministic, which may be unexpected to the user. And the more-explicit function to convert steady timestamp to string formatted it as "1:23:45.678 ago" rather than the usual "2019-11-12 08:26:42.123", to make it more obvious that the result may change from call to call. But ever since I wrote that code, it probably hasn't been useful once, only added some minor annoyance. So this diff gets rid of all that and just adds toString() method to all Timestamp-s.

Format for steady_clock Timestamp-s changes from "1:23:45.678 ago" to "2019-11-12 08:26:42.123". But afaict none of the few call sites care about it. I don't remember ever seeing the "X ago" thing in logdeviced's logs. (Well, I've seen it once, and it was broken - the timestamp was in future, and unsigned subtraction overflowed and printed something like "0:00:18446744073709551600.18446744073709551602 ago", and that's what prompted me to do this diff.)

Reviewed By: scsiguy

Differential Revision: D18434087

fbshipit-source-id: 03c7a5828d95afbaaaa23b808085ac70180f534f