Remove HealthMonitor polling from FailureDetector
Summary:
Removed `HealthMonitor` polling from `FailureDetector` to get node status. Instead `HealthMonitor` updates a field in `FailureDetector` with the current status.
Unless `FailureDetector` is explicitly assigned to `HealthMonitor` there is no interaction between the two.

In `Server` in `initFailureDetector()` if failure detector is created then it is assigned to `HealthMonitor`, if it was previously created via `initProcessor()`.

Reviewed By: mcrnic

Differential Revision: D18173769

fbshipit-source-id: 3a56aa2449cfe0faaae9d2aa5ad9965e60f1598f